coins: move SendAmount, TxValidationError out btc
Move out of the btc package, make available for all coins.
